Rename app > app-core, Introduce GenericApp
