+++ pom.xml
@@ -53,8 +53,8 @@
 
     <modules>
         <module>annotations</module>
-        <module>app</module>
-        <module>app-test</module>
+        <module>app-core</module>
+        <!--<module>app-test</module>-->
         <module>bom</module>
         <module>config</module>
         <module>config-test</module>
@@ -99,7 +99,7 @@
             </dependency>
             <dependency>
                 <groupId>net.spals.appbuilder</groupId>
-                <artifactId>spals-appbuilder-app</artifactId>
+                <artifactId>spals-appbuilder-app-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
